requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
loguru==0.7.2
multidict==6.0.4
numpy==1.26.0
openai==0.28.0
pycparser==2.21
pydub==0.25.1
pyperclip==1.8.2
python-dotenv==1.0.0
requests==2.31.0
simplejson==3.19.1
sounddevice==0.4.6
soundfile==0.12.1
tqdm==4.66.1
urllib3==2.0.5
yarl==1.9.2
